seguranca/pyproject.toml
[project]
name = "seguranca"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aes-python>=1.5.5",
    "flask>=3.1.3",
    "galois>=0.4.10",
    "numpy>=2.4.3",
    "pycryptodome>=3.23.0",
]
